+++ ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: gd-harco <gd-harco@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/11 14:07:51 by gd-harco          #+#    #+#             */
-/*   Updated: 2022/11/14 13:40:06 by gd-harco         ###   ########lyon.fr   */
+/*   Updated: 2022/11/17 21:44:05 by gd-harco         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	int	x;
+	size_t	x;
 
 	x = 0;
 	while (n > 0)
